Refactor tsconfig resolution logic into utility function
