+++ libs/nx/src/index.ts
@@ -1,1 +1,2 @@
+export * from './lib/get-ts-config-path';
 export * from './lib/project-root';
