+++ libs/tsc/src/executors/typecheck/executor.ts
@@ -1,7 +1,6 @@
-import { existsSync } from 'node:fs';
 import { resolve } from 'node:path';
 import { promiseSpawn } from '@container/docker';
-import { projectRoot } from '@container/nx';
+import { getTsConfigPath, projectRoot } from '@container/nx';
 import type { Executor } from '@nx/devkit';
 import type { TscTypecheckExecutorSchema } from './schema';
 
@@ -9,28 +8,13 @@ export interface TypecheckExecutorOptions {
   tsConfig: string;
 }
 
-const possibleTsConfigs = [
-  'tsconfig.spec.json',
-  'tsconfig.base.json',
-  'tsconfig.json',
-  'tsconfig.app.json',
-  'tsconfig.lib.json',
-];
-
 const tscTypecheckExecutor: Executor<TscTypecheckExecutorSchema> = async (
   { tsconfig = 'tsconfig.spec.json' },
   context,
 ) => {
   const root = projectRoot(context);
 
-  if (existsSync(resolve(root, tsconfig))) {
-    for (const config in possibleTsConfigs) {
-      if (existsSync(resolve(root, config))) {
-        tsconfig = config;
-        break;
-      }
-    }
-  }
+  tsconfig = getTsConfigPath(tsconfig, root);
   try {
     await promiseSpawn('node_modules/.bin/tsc', ['--noEmit', '-p', resolve(root, tsconfig)], {
       cwd: context.root,
